Moved AesEncryptionManager to shared module
